% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAndUnzipShp.R
\name{downloadAndUnzipShp}
\alias{downloadAndUnzipShp}
\title{A Function to download and unzip a shapefile into a temp folder}
\usage{
downloadAndUnzipShp(url)
}
\arguments{
\item{url}{to download the data.}
}
\description{
This function helps to manipulate the data available at GeoSampa website.
}
\examples{
downloadAndUnzipShp("http://geosampa.prefeitura.sp.gov.br/PaginasPublicas/downloadArquivoOL.aspx?orig=DownloadCamadas&arq=03_Equipamentos\%5C\%5CAbastecimento\%5C\%5CShapefile\%5C\%5Cequipamento_abastecimento&arqTipo=Shapefile")
}
\keyword{url}

